refactor: move dry_run_scaffold to local.py
