+++ scaffoldr/local.py
@@ -53,3 +53,38 @@ def scaffold(
     _git(["commit", "-m", "chore: initial scaffold"], cwd=root)
 
     progress(f"Done. Project ready at {root}")
+
+
+def dry_run_scaffold(
+    project_name: str,
+    template_name: str,
+    path: Path,
+    progress: Callable[[str], None] = lambda _: None,
+) -> None:
+    cfg = Config.load()
+
+    root = path / project_name
+
+    if root.exists():
+        raise LocalError(f"Error: {root} already exists.")
+
+    scaffold_variables = build_scaffold_variables(
+        project_name, cfg
+    )
+
+    template_path = resolve_template_path(template_name)
+    template: Template = load_template(template_path)
+    rendered_template = render_template(
+        template, scaffold_variables
+    )
+
+    for path_str, _ in rendered_template.items():
+        progress(f"[dry-run] Would create: {(root / path_str)}")
+
+    # git
+    progress(f"[dry-run] @{root}: git init")
+    progress(f"[dry-run] @{root}: git add .")
+    progress(
+        f'[dry-run] @{root}: git commit -m "chore: '
+        'initial scaffold"'
+    )
